net: core: Optionally run link watch event on highpri wq
On systems where reaction time to link events is critical, offer a
config option to move the link event processing to the high-priority
workqueue.

This is useful on networking devices where link events might trigger
reconfigurations in other parts of the network (STP/RSTP, SPB, ring
protocols etc.), and the time that parts of the network is unavailble
must be kept to an absolute minimum.