net: bridge: Add support for software port in locked mode
In a 802.1X scenario, clients connected to a bridge port shall not
be allowed to have traffic forwarded until fully authenticated,
in which instance a static fdb entry of the clients MAC address is
added to he bridge, so as to allow for the client to have it's
traffic forwarded.
This scenario is facilitated with setting the bridge port in locked
mode, and is supported by various switchcore chipsets.