net: bridge: add control of bum flooding to bridge itself
The bridge itself is in many ways also a port, but it does not have a
'struct net_bridge_port'.  In many cases we want to treat it as a proper
port so concessions have been made, e.g., NULL port or host_joined
attributes.

This patch addresses another such case by adding support for controlling
flooding of unknown broadcast/unicast/multicast to the bridge itself.
Not only does this give us another knob to control what reaches upper
interfaces, it also allows us to use this one source of information for
controlling what is flooded to the CPU in an offloaded switching
fabric.  (Promisc and foreign ports are other sources we want to query.)

The fast path is invaded with a few more if-statements.  This could
possibly be addressed by speculatively setting local_recv = enum; and
postpone the decsion to br_pass_frame_up(), this is a lot more work.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>