net: sparx5: Add procfs debug code from Microchip
This is some old code for debug info through procfs
that I was given by Microchip.  They do not want this
upstreamed and would like to use debugfs in the future.
But it's good that we at least have something for now.

Information is available under /proc/sparx5_*
It has ports, cpu ports, portstate, mactable, and fdmastate.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>