net: dsa: Add assisted multicast router port mode
For hardware that does not have a separate way of including multicast
router ports in the forwarding set of registered IP multicast groups,
add an opt-in mode where drivers can request that the DSA layer will
fold mrouter ports into existing MDB entries. Note: drivers are still
responsible for ensuring delivery of unregistered multicast.

It will also keep track of the fact that on any given port, a group
may be added to the hardware entry because (a) the port is a member of
the group, (b) there are other subscribers to the group and this port
is an mrouter port, or (c) all of the above.

On delete this makes things more complicated, as we must ensure that
neither (a) nor (b) is true before removing a port from the hardware
MDB. Furthermore, we must ensure that when the last subscriber leaves
a group, the whole entry is removed, such that we don't keep entries
with _only_ mrouter ports in the hardware MDB.

As an example, say that you have three offloaded MDB entries with some
subscribers:

                   Physical ports
Address            0  1  2  3  4  5
01:00:5e:00:00:01  X  .  .  .  X  .
33:33:00:00:00:02  .  X  .  .  X  .
01:00:de:ad:00:03  .  .  .  X  .  .

If, at this point, port 2 is marked as a mrouter port, the folded MDB
mode will add port 2 to all existing entries (and update any
intermediate DSA ports, if needed) to produce the following:

                   Physical ports
Address            0  1  2  3  4  5
01:00:5e:00:00:01  X  .  X  .  X  .
33:33:00:00:00:02  .  X  X  .  X  .
01:00:de:ad:00:03  .  .  .  X  .  .

Note that the final entry is not affected, as it is not an IP
multicast group.

Let's say that both ports 1 and 4 leaves IPv6 group 2. At this point,
the hardware entry only contains mrouter ports, and will therefore be
removed. At that point, whatever mechanism the driver uses to flood
unregistered multicast will take over to make sure that port 2 still
receives the group's messages.

                   Physical ports
Address            0  1  2  3  4  5
01:00:5e:00:00:01  X  .  X  .  X  .
01:00:de:ad:00:03  .  .  .  X  .  .

Now let's first say that an IGMP report for 239.0.0.1 is received on
port 2. This will not require any hardware operation as the port is
already marked as a member due to its mrouter port status. Then
suppose that port 2 loses its mrouter port status. At this point the
folded MDB mode will only remove port 2 from IPv6 group 2, since port
2 is still a member of group IPv4 group 1:

                   Physical ports
Address            0  1  2  3  4  5
01:00:5e:00:00:01  X  .  X  .  X  .
01:00:de:ad:00:03  .  .  .  X  .  .