spi: spi-drg: Add driver
PONCat3 has a small companion core (ARMv5) called "Dragonite",
intended to serve as a dedicated PoE controller. On Dagger boards,
this core runs a custom firmware that exposes its SPI controller, to
which the system's FLASH is attached, to the main core running Linux.

This commit adds a driver for that implements the host side of that
interface.

This patch is copied from WeOS repo, #acd0bb3.

Signed-off-by: Anders Öhlander <anders.ohlander@westermo.com>