net: bridge: mrp: Add support for VID in MRP Frames
This patch adds support for configuring VID to be used and expected
for a specific ring instance when it sends and receives MRP frames.

The configuration of the VID is tied to a specific MRP instance and is
configured via netlink like other MRP settings.

If a instance is set to use a VID the test frames that are generated
will insert the VID in the frames. The VLAN for the provided VID will
be created and the ring ports added to it.

If an instance is configured with a VID the incoming frames will be
expected to have a VLAN tag. Otherwise those frames will be
discarded.

If no VID is provided it will be 0, i.e no VID is expected for the MRP
frames.