net: dsa: mv88e6xxx: Fix false positive busy bit polling timeouts
Rework the polling logic to include a final sampling attempt after the
timeout, such that we are guaranteed to give the hardware at least 50
ms to carry out the operation.

Previously, the busy bit polling logic would terminate when the
timeout had passed and the bit had been polled at least twice. Since
the logical operation will result in multiple MDIO operations on a
shared bus, blocking may occur after the actual sampling of the bit
took place. I.e. a typical indirect read will look like this:

1. Write command register to trigger read
2. Read command register to verify that data register is valid
3. Read data register to fetch register value

The bit is sampled at some point after 1 and before 3. This means that
we can end up with this sequence of events:

|  Time | Op     | Comment                              |
|-------|--------|--------------------------------------|
|  0 us | W cmd  | Busy bit is sampled for the 1st time |
| 20 us | R cmd  |                                      |
| 40 us | R data |                                      |
| 60 us | W cmd  | Busy bit is sampled for the 2nd time |
| 80 us | R cmd  | At this point our process is blocked |
| 67 ms | R data |                                      |

At this point the loop will terminate, as we've sampled the bit twice
and our time is up. But we've actually only given the hardware 60 us
to perform the operation.