net: bridge: mrp: When changing port state notify br port state selection
This may be a workaround for a larger issue, how MRP correctly should
interact with the necessary parts of the stp code.

Currently the way MRP changes port states and how it interacts with
the stp code, will cause the carrier on the bridge interface to not be
changed correctly.

This patch adds a call to the br_port_state_selection() function when
mrp changes the state on one of its ports. We need to do this in order
to see if we should set carrier on or off on the associated
bridge. Otherwise we can end up in a state where you have ports that
are up but the bridge will still have carrier down.