net: bridge: mrp: On frame rcv as MRC set correct hw_domain
When we receive mrp frames when acting as a MRC, it is possible that
we either want to forward the frame even though we may have some
offloading.

An example of this can be when you have some situation where the HW
cannot forward frames, even though it has some offloading. Another
case could be where you have one ring port that is not offloadable,
e.g. a ssl port.

Therefore we want to ensure that the offload_fwd_mark will be handled
correctly before we call br_forward.

We do this by ensuring that the received skb belongs to the same
hw_domain as the target port, by calling nbp_switchdev_frame_mark().

This will have either of the following effects, depending on the
current state of the offload_fwd_mark:

  1. If offload_fwd_mark=0, i.e. the frame has not been forwarded in
     hw (no tx fwd offloading of MRP frames). In this situation the
     frame should not have its offload_fwd_mark be changed to 1 during
     the br_forward call chain in nbp_switchdev_can_offload_tx_fwd().

  2. If offload_fwd_mark=1, i.e. the frame has been forwarded in
     hw. Then when we reach the br_forward call we want that to notice
     that the frame should not be forwarded, in the call to
     nbp_switchdev_allowed_egress(). We need to ensure the same
     hw_domain and offload_fwd_mark is 1.

Note, this should only have any sort of effect when you have some
underlying HW. If we are fully in software this should not matter.