net: bridge: mrp: Improve verification of HW support on switchdev calls
If general if a switchdev call returns a EOPNOTSUPP, we will assume
that we are able to run in SW. We will therefore not return that as an
error that will stop MRP from being configured.

If this is not done it will not be possible to run any image on a
virtual setup, if the image has been built with CONFIG_NET_SWITCHDEV
as part of the kernel config.