net: dsa: mv88e6xxx: Add initial support for configuring MRP
This commit adds basic support for configuring the HW to be able to
receive MRP frames.

The extent of the support entails adding correct MGMT entries to the
ATU for the MRP MAC addresses:

  - test frames: 01:15:4e:00:00:01
  - control frames: 01:15:4e:00:00:02

Currently we will always add both of the MAC addresses with the
upstream port as destination. This ensures that we will get the frames
to the CPU, for both MRM and MRC.

If we are configured to act as a MRC we can also offload the sending
of MRP frames to the other ring port. This is done by simply adding
both of the ring ports to the ATU, in addition to the upstream
port. Note that this will only work when both the ports are located on
the same switch device. If they are not, the frames need to be
forwarded from the SW bridge in br_mrp, after it has been received
towards the CPU.

It is important that it is ensured the offload_fwd_mark is set
correctly when frames are received in the DSA tag handling. If it is
not handled correctly it is possible that we may forward duplicate
frames when we have the possibility to offload the sending of frames
between the ring ports.

The reason all of the ATU entries have to be MGMT, is so that we can
receive the frames through blocked ports. This is important regardless
if you act as a MRM or MRC.