net: dsa: mv88e6xxx: Add initial suport for tagger_data
This commit adds inital support for sending tagger_data between the
switch to the tagger.

In this inital commit the only infomration that can be passed is an
indication if a port has had the possibility to offload forwarding of
MRP frames.

We need to know this in order to assign the offload_fwd_mark correctly
on any potential incoming MRP frame. If the tagger infomration
indicates that a port will have forwarded the frame in HW we must mark
the offload_fwd_mark to 1. Otherwise the frame will be forwarded again
from the SW and cause duplication of the MRP frames.

Therefore in summary, we check any traped frame if it is also an mrp
frame. If this is the case, check the tagger_data inforamtion with the
index of the received port and switch. If the port indicates it has
mrp tx fwd offloading, we do nothing, otherwise change the
offload_fwd_mark=1.