net: bridge: switchdev: Add func to check if br port is offloaded
This commit adds a very simple function that simply returns a boolean
response if a provided bridge port has any switchdev offloading.

The function will check this by simply referring to the offload_count
variable, as long as it is zero we have no presumed offloading
otherwise we do.

If no switchdev support is compiled for the kernel build, this
function will always return false.