net: bridge: mrp: Check if both ring ports are offloaded
This commit adds a check if both of the configured ring ports are
offloaded by switchdev. If any of the ports are not offloaded we will
specify that we run MRP in SW, even if one of the ports could still be
offloadable, in that case we perform as much offloading as possible.

One case where this is a relevant check to perform is when you have
one ring port that is an offloaded port and another that is not,
e.g. a ssl port.

In this case we still need to possibly perform some offloading of the
port that has underlying HW, so that we can ensure that we receive any
MRP frames to the SW. Afterwards the MRP handling will be performed in
SW.

Therefore, the check is simply if one of the ports is not offloaded,
then we treat it as running MRP in SW.