net: phy: Optionally run state machine on highpri wq
On systems where reaction time to link events is critical, offer a
config option to move the state machine processing to the high-priority
workqueue.

This is useful on networking equipment where link events might trigger
reconfigurations in other parts of the network (STP/RSTP, SPB, ring
protocols etc.), and the time that parts of the network are unavailble
must be kept to an absolute minimum.