net: sparx5: Fix extra VLAN tag being added when no untagged VLAN
When a port does not belong to a untagged VLAN it was set to tag all
frames. This meant it added a tag with vid 0, even if the frame was
already tagged. Change it so it does not use the "tag all" setting.

This only affected frames being transmitted from the kernel. Offloaded
switching did not have this issue and does not appear to be affected by
this change.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>