net: dsa: mv88e6xxx: Handle wraparound for RMON counters
Some rmon counters in the mv88e6xxx family is 32 bits long,
which will mean they could wrap around in a short amount
of time. Use a carry array to extend them to 64 bits.
This means also that we must read them periodic to see
if they have wrapped around.

Signed-off-by: Mattias Forsblad <mattias.forsblad@westermo.com>