net: dsa: mv88e6xxx: Busy-wait for busy bits [INTERNAL]
Upstream code will, sensibly, fallback to a sleeping wait for busy
bits if the bit is still set after the first two busy polls.

However, if this happens as the system is discovering that a link went
down, and that link happens to be an active ring port, it has a direct
impact on the whole network's convergence time.

Therefore, it makes more sense for us to just keep hammering the bit
until it clears (which should just be a few more retries at the most
anyway).