mtd: spi-nor: Fix sector erase on basis/envoy platforms
The current driver for the mtd/spi-nor flash on the basis and
envoy platform does not work in 4-4-4 quad-spi mode.
Setting it back to do regular 1-1-1 reg command fixes the problem
temporarly on those platforms. This fix will be reverted as soon
as the driver is fixed.

Patch brought in from Netbox.