net: dsa: mv88e6xxx: Refactor PHY IRQ handling
This is a rework of the PHY interrupt handling based on our previous
patch: 440aaede3f7372f9502bf4ff56072cf686b8ef98.

The essence of the previous patch is mostly the same. Some
re-structuring have been made (for the better?).

The main issue that the patch aims to resolve is the general handling
of the PHY interrupts, since the way this is done differs on older
chipsets, e.g. 6097. On the newer chipsets, what the upstream
implemenation consider, the interrupt mask register in Global 2
includes the specific ports directly. Comparatively, the older
chipsets need to look into a separate register to mask what ports have
received an interrupt.

The main difference that this refactoring have to our previous is that
the intention is for it to be somewhat less intrusive on existing IRQ
set up handling. Like before, a seprate IRQ object is defined the be
specifically responsible for the PHY interrupts, in addition to the
existing Global 1 and Global 2 interrupt handlers.

In the chipset info struct a boolean is added that can be used to
indicate that the specific chipset uses the older Global 1 PHY
interrupt handling. If this is set, the interrupt handling will be set
up in the manner that allows for that case to function.

In the other case, it should still work in the same manner as it does
upstream, except that the PHY interrupts will still be specified in
the specific PHY IRQ object, instead of the Global 2 IRQ handler.

In our previous patch we specified the number of PHY IRQs in the
infomration of the chip, however we know simply rely on the existing
definition on the number of internal PHYs that exist. We assume that
we can configure the interrupts, if the Global 1 and Global 2 IRQs can
be configured, since if those exist the PHY IRQ handling should be
possible.

Signed-off-by: Jonas Holmberg <jonas.holmberg@westermo.com>