net: bridge: mrp: Allow EAGAIN as valid error code to use sw_backup
When MRP tries to configure the MRP ring role in the HW using
switchdev calls it has two steps:

  1) Try to configure "Fully" in hw. If this works we are done.
  2) If we could not configure "Fully" in hw try to do it partially by
  setting the sw_backup variable in the switchdev call.

In order to try again the retrun value of the switchdev_port_obj_add()
call must currently be EOPNOTSUPP. However when an EOPNOTSUPP is
attempted to be returned from the driver level, the dsa layer will
capture any EOPNOTSUPP return value and in turn return it as ok (0),
upward thorugh the stack. Hence, when the call reaches back to the mrp
switchdev part it is not returned as EOPNOTSUPP.

This patch simply adds the possiblity to also return with EAGAIN as
the error code from the lower driver level.