net: bridge: mrp: Reset test_count_miss on state change to close
When the ring state changes from open to closed we want to set the
test_count_miss variable back to 0. Doing this should be inline with
the specification.

Most of the time this will not be an issue, but an example where this
can be a problem is as follows.

  1) The ring is currently open and is broken on the MRM. The MRM is
     in this case in the PRM_UP state.
  2) When the link goes up the MRM will instantly move to the Closed
     state from this position, without waiting for any ring frames to
     traverse the ring.
  3) In this case if the miss count is not reset, when the miss check
     worker is invoked for its next iteration it will instantly move
     the ring back into the open state, because the counter indicates
     that we have missed sufficient frames, since it has not been set
     back to 0.
  4) When this occurs, the ring will go back to open and therefore set
     both of the ring ports in forwarding.
  5) Now it will only go back to closed when the ring frames are
     received.

This situation should potentially be able to open up a small timing
window where both of the ring ports on the MRM will be in forwarding
when the ring is closed, thus allowing a broadcast storm to possibly
happen.

Therefore, simply do what seems to be indicated by the specification
and set the miss counter to 0 when the ring state changes to closed.