net: dsa: mv88e6xxx: tc: Set MAC qpri and fpri.
Adapt the DSA flower handlers to support configuration of the
MAC qpri and fpri for MC addresses.

This is an example about how to set the MAC qpri and fpri to 7
for MC MAC address 01:80:c2:00:00:10 as management frame.

tc filter add dev ethX3 ingress protocol 802.1Q prio 700 \
	flower \
	dst_mac 01:80:c2:00:00:10 \
	action vlan push protocol 802.1Q priority 7 id 1 \
        action trap

The 'action trap' should not be specified if this is just a
priority override only.

Known limitations:
- The MAC qpri and fpri cannot be set individually.
- This should be executed on at least one port of every switchcore.

Signed-off-by: Lucian Banu <klucba@westermo.com>