net: bridge: Carrier down on bridge when all ports are disabled
If all bridge ports are in disabled state the bridge device should be
down otherwise it should be up.
The bridge ports states are disabled, listening, learning, blocking
and forwarding.

In earlier code at least one port had to be in forwarding state for
the bridge interface to be up.
This could cause problem when using a redundancy protocol when one
port was in forwarding and the other were in blocking, backup ports.
If the forwarding port had link down then one of the backup ports
should transfer state to forwarding. Because all ports are now in
blocking state the bridge device changed state to down until one of
the backup ports changed state to forwarding then the bridge interface
came up again.
The interfaces above using the bridge interface will also have a
change in their state from up to down and then up again. This could
result in loss of communication for applications using those
interfaces.

To upstream this patch a setting for this behavior i probably needed
so both the previous and new behavior can be chosen.

Signed-off-by: Greger Wrang <greger.wrang@westermo.com>