net: bridge: mrp: On port state change also set MST state, if enabled
This patch is another workaround because of the fact the the port
MRP_AWARE condition exist in the STP part of the bridge. Because of
this we will not be allowed to reach the 'br_mst_set_state()' function
call in the STP part of the code before we just exit.

Therefore, since MRP wants to totally control the port states of the
ports, without any interference from STP, we need to signal the mst
state when MRP changes the port states.