net: dsa: mv88e6xxx: Fixed LAG link issue when LAG is under bridge
The LAG link remains down when it is under br0, because the DSA
ports are disabled due to inheritance of lag device state.

It's looking like a cycle dependency between LAG and DSA ports state.
The LAG needs the LACP traffic to passthrough the ports
before change its state, but the ports are disabled due to inheritance.

The change avoid the change state of DSA port while joining the bridge
if this is part of a LAG.
The DSA port is already set in FORWARDING state when its added to
the LAG by team_add_slave()->dev_open().

Signed-off-by: Lucian Banu <klucba@westermo.com>