dsa: move the functions checking if faster ageing time on same switch.
Move the function used for checking if a faster ageing time is set on
an other bridge on the same switchcore. We only want to change to a faster
ageing time if an other bridge have a lower value, not if an port in the own
bridge since all ports not handle yet will have a lower value.
This must be done in the port layer before calling the switch layer
since then there is knowledge of what bridge the changed ageing time
originated from.