net: dsa: Include host MDB entries in assisted mrouter port cache
Make sure that any group memberships on the local bridge are added to
the corresponding folded MDB cache entry, because if the cache is
bypassed any external queriers will not be included in the group's
hardware entry. E.g. in a scenario like this:

      .--- Local subscriber
      o
     br0
     / \
  swp0 swp1
        o
        '--- Querier

When the local subscriber registers its membership, we must ensure
that swp1 is also included in the resulting hardware entry - as it is
marked as a multicast router port.