net: dsa: mv88e6xxx: rmu: Use RMU for dumping the ATU.
The dump of the ATU over MDIO is very slow and scales bad with
increasing number of ATU entries. Instead use the RMU dump ATU
command to efficiently retrieve the whole ATU DB.
Because of the calls are per port while one RMU dump command
dumps the whole ATU, hold on to previous response for some
time, caching the already received data.

Signed-off-by: Mattias Forsblad <mattias.forsblad@westermo.com>