net: switchdev: bridge: Avoid handling local IPv4 MC in L2 ATU
Many services (IGMP, VRRP, OSPF etc.) rely on IPv4 multicast in the local
range (224.0.0.x). This range overlaps on L2 level with 224.0.0.x -
239.0.0.x and 224.128.0.x - 239.128.0.x. This patch blocks any attempt
to add/remove any ATU entry in these ranges through the switchdev layer,
so that local IPv4 multicast is guaranteed to be flooded according to
flooding settings.

Signed-off-by: Jacques de Laval <Jacques.De.Laval@westermo.com>