switchdev: Always use defer when calling driver
When many calls are done through switchdev and some are using
SWITCHDEV_F_DEFER and some do not the calls down to the driver can be
reordered. Calls may fail if a call is dependent on previous.

For example adding or removing a mdb address to a vlan. To add a mdb
address the vlan must exist otherwise the mdb add will fail. The
opposite it true when the mdb address should be removed.  This is true
when calling the dsa driver but this is because the bridge and the dsa
are handling the vlan and fdb/mdb differently. Another solution and
possible the best one is to change the behavior so the dsa driver
works in the same way as the bridge. For now the solution with defer
is chosen as it is easier and do not have any known drawbacks.