dsa: Propagate DSA header priority properly on ingress and egress.
We want to propagate the packet priority to/from the switchcore
for correct prioritization in the switchcore.

On ingress, if the packet is untagged, create a priority tagged 802.1Q
packet with the DSA priority if the port is member of a vlan filtering
bridge. If the port isn't a member of a vlan filtering bridge,
strip the DSA header.

On egress, if the packet is untagged, set the skb priority
in the DSA header. This implies that the socket has set
priority via the SO_PRIORITY setsockopt.

In either case if the packet is already 802.1Q tagged, the
priority should already be set in that tag, which will be
transferred to the DSA tag.

Signed-off-by: Mattias Forsblad <mattias.forsblad@westermo.com>