driver: net: dsa: mv88e6xxx: fixed purgeing locked entries when ageing out.
The compare of mac addresses in 'mv88e6xxx_atu_locked_entry_find_purge'
will now purge the address from the internal locked list if equal to address
provided in function parameter.
The entry in the internal locked list will be purged when a new entry
with that mac is added from user.

The result of the previous code was the the internal locked entry was not
removed and when that aged out it removed the entry added by the user, eg.
the authed entry, and a new authentication was needed.