net: bridge: br_fdb: delete from fdb also if added by user as dynamic entry.
When a entry age out from the atu that is added by user as a dynamic entry
it will not be removed from bridge fdb since it is no externally learned.
Now the entry will be removed if it is externally learned or if it is
added by user and the offloaded flag is set. This is the way locked entries are added
or replaced when a mac addres is authenticated by either mac-auth or dot1x.