driver: net: dsa: mv88e6xxx: Add function to do load_purge on portvec
Currently the mv88e6xxx_port_db_load_purge() function can only be
provided with a single port. In some cases, especially when adding
multicast entries that should be added to several ports at the same
time, it is more efficient to allow the user to provide the entire
port vector, instead of doing one port at a time.

This commit changes the load_purge function to use a portvec input
value instead of just a port. The old function now acts as a base
function that is called by the two new wrapper functions:

  * mv88e6xxx_port_db_load_purge()

    * This still behaves in the same way as today, taking in a
      port. This port is translated to a portvec and passed to the
      load_purge base function.

  * mv88e6xxx_portvec_db_load_purge()

    * This is the new function that takes a portvec and it simply
      passes it to the load_purge base function. Technically the base
      load_purge function could be called directly, but this is done
      for the sake of clarity and a hopefully more understandable
      API.

This should reduce the risk to introduce problems in existing code,
since any function that calls mv88e6xxx_port_db_load_purge() will
still do so, it is only translated to a portvec with one port inside.

It also introduces a check in the base load_purge function that checks
if the portvec only contains a single port when a Unicast address is
attempted to be added. Probably this check could be okay to do, since
you most likely will never want to provide a range of ports when
dealing with a Unicast address.

Signed-off-by Jonas Holmberg <jonas.holmberg@westermo.com>