net: bridge: switchdev: Only notify switchdev when a port group is added
When a port group is added for a multicast group the switchdev layer is
notified through it's subscription on MDB events. If the port group is
later modified, eg. if the source list is modified, another RTM_NEWMDB
notification may be sent. This is not interesting for the switchdev layer
as it only cares about the destination port. Furthermore: more than one
notification about new MDB entries for the same port and group can mess up
reference counting done in lower layers.

This patch adds a flag to keep track of if the switchdev layer has already
been notified of a destination port added, and thus avoids the problems
described above.

Signed-off-by: Jacques de Laval <Jacques.De.Laval@westermo.com>