sparx5: Set mac address per port based on used ports in dts
Microchips previous solution was to base the mac address offset per
port based on the physical connection on the SC. This means that if
you don't use the first ports on the SC the offset from base_mac will
be greater than anticipated. This solution starts with the first port
defined in the dts instead.

Signed-off-by: Lennart Eriksson <lennart.eriksson@westermo.com>