net: bridge: mrp: Clear the BR_MRP_LOST_CONT bit from the port flags
This patch adds two situations where we ensure we clear the
BR_MRP_LOST_CONT port flags bit.

  1: When we go from ring state Open -> Closed, sent from user space.

  2: When we receive ring frames as MRM, we also always ensure that
     the flag is cleared.

The reason we do this is because otherwise a possibility exist where we
can end up in closed state but go back to open because the user space
part of the MRP handling reads the IFLA_PORTINFO containing the
IFLA_BRPORT_MRP_RING_OPEN value, which gets its value based on the
flag BR_MRP_LOST_CONT.

This has specifically been observed when the MRM is in the PRM_UP
state and its secondary ring port gets link up, and the ring will be
fully closed, no other breaks in any part of the ring. In this
scenario we should go directly to Closed and stay there. But because
of this issue a specific timing could occur where we go back to Open
and then back to closed again when we receive a ring frame.

This is a very specific corner case that seem to require a very
specific timing and bad luck for it to occur, but nevertheless it
could. In the worst case scenario when this happens, it is possible
for a broadcast storm to occur, since all ports may be in full
forwarding during a very short period.

Signed-off-by: Jonas Holmberg <jonas.holmberg@westermo.com>