fixup: net: bridge: mrp: Add support for VID in MRP Frames
During porting of our MRP VID handling from the 4.19 kernel, one
adjustment that was intended and described as being part of this
original commit, was missed. This commit simply includes a line that
should have been part of this from the beginning.

For context on this change, why and when it was introduced on the 4.19
kernel, refer to its original commit:

  - 5f4a2781dc160dad45e346670561e5dce20b2317

A short summary of this adjustment is that we only care to check if
any potential VID in the MRP frame is correct, if the MRP instance
have been configured to expect any specific VID. Otherwise we do not
care to check.

Checklist WQMS-355909421-1808, rev 1.0, Verified by Fredrik Pettersson,
Review participants Jonas Holmberg