net: Introduce IFF_URGENT flag
Some interfaces events are time sensitive, for example
interfaces that run ring protocols. The link up and down
events needs to be propagated quickly. The standard implementation
defers down events to next scheduled run of the link watcher.
Make it possible to set a interface flag to urgent and
check for that flag in the urgent classification routine thus
scheduling that event immediately.

Signed-off-by: Mattias Forsblad <mattias.forsblad@westermo.com>